refactor: all audio infos components;misc: open umi's MFSU
