+++ src/components/AudioInfos/index.tsx
@@ -1,0 +1,15 @@
+import powerplant from './powerplant';
+import propeller from './propeller';
+import signalInfo from './signalInfo';
+import typeRadio from './typeRadio';
+import radiationTarget from './target/radiationTarget';
+import echoTarget from './target/echoTarget';
+import pulseTarget from './target/pulseTarget';
+
+export const Powerplant = powerplant;
+export const Propeller = propeller;
+export const SignalInfo = signalInfo;
+export const TypeRadio = typeRadio;
+export const RadiationTarget = radiationTarget;
+export const EchoTarget = echoTarget;
+export const PulseTarget = pulseTarget;
