+++ src/pages/audioImport/AddSound.tsx
@@ -1,33 +1,16 @@
 import React, { useState, useEffect } from 'react';
-import { connect, Dispatch, history } from 'umi';
-import {
-  DatePicker,
-  TimePicker,
-  Result,
-  Row,
-  Col,
-  Input,
-  Button,
-  Select,
-  InputNumber,
-  Upload,
-  message,
-  Modal,
-  Form,
-  Radio,
-  Steps,
-} from 'antd';
-import { InboxOutlined } from '@ant-design/icons';
+import { connect } from 'umi';
+import { Row, Col, Button, Form } from 'antd';
 import moment from 'moment';
-import Cookies from 'js-cookie';
-import style from './style.less';
-import CookieUtil from '@/utils/cookie.js';
-import RecorderCn from '@/components/recorder_v2/index.jsx';
-import { SERVICEURL } from '@/utils/const';
-
-const { Dragger } = Upload;
-const { Option } = Select;
-const { Step } = Steps;
+import {
+  TypeRadio,
+  RadiationTarget,
+  EchoTarget,
+  PulseTarget,
+  Powerplant,
+  Propeller,
+  SignalInfo,
+} from '@/components/AudioInfos/index';
 
 // TODO 现仅将组件AddSound抽象出来，这里由于不了解需求，暂搁置优化
 const AddSound: React.FC<{
@@ -96,753 +79,11 @@ const AddSound: React.FC<{
     }
   }, [InforImport.searchInfor]);
 
-  // 音频类型单选框
-  const TypeRadio = () => {
-    const onChange = (e) => {
-      // type = e.target.value;
-      settype(e.target.value);
-      console.log('radio checked', e.target);
-      // console.log(sumForm.getFieldsValue());
-    };
+  const modifyActionNames = ['modifyNoise', 'modifyEcho', 'modifyPulse'];
 
-    return (
-      <>
-        <Form.Item
-          name="signal_type"
-          label="信号类型"
-          labelAlign="left"
-          labelCol={{ span: 2 }}
-        >
-          <Radio.Group onChange={onChange} value={type}>
-            <Radio value={1}>辐射噪声</Radio>
-            <Radio value={2}>目标回声</Radio>
-            <Radio value={3}>主动脉冲</Radio>
-          </Radio.Group>
-        </Form.Item>
-      </>
-    );
-  };
-
-  // 辐射噪声特有信息表单
-  const RadiationTarget = () => {
-    const [value_1, setValue_1] = useState(-1);
-    const [value_2, setValue_2] = useState(-1);
-    const [visible, setVisible] = useState(false);
-
-    const [form] = Form.useForm();
-
-    const onChange_1 = (e) => {
-      console.log('RadiationTarget checked', e.target);
-      setValue_1(e.target.value);
-    };
-    const onChange_2 = (e) => {
-      console.log('RadiationTarget checked', e.target);
-      setValue_2(e.target.value);
-      if (e.target.value === '添加新类别') {
-        setVisible(true);
-      }
-    };
-
-    return (
-      <>
-        <Row gutter={16}>
-          <Col span={24}>
-            <Form.Item
-              name="is_over_water"
-              label="水面/水下"
-              labelAlign="left"
-              labelCol={{ span: 2 }}
-            >
-              <Radio.Group onChange={onChange_1} value={value_1}>
-                <Radio value={1}>水面</Radio>
-                <Radio value={0}>水下</Radio>
-              </Radio.Group>
-            </Form.Item>
-          </Col>
-        </Row>
-        <Row gutter={16}>
-          <Col span={24}>
-            <Form.Item
-              name="rn_type"
-              label="目标类型"
-              labelAlign="left"
-              labelCol={{ span: 2 }}
-            >
-              <Radio.Group onChange={onChange_2} value={value_2}>
-                {InforImport.rnType?.map((item) => {
-                  return <Radio value={item.rn_type}>{item.rn_type}</Radio>;
-                })}
-              </Radio.Group>
-              <Button
-                value={'添加新类别'}
-                onClick={() => {
-                  setVisible(true);
-                }}
-                style={{
-                  display:
-                    CookieUtil.get('role') === '3' ? 'none' : 'inline-block',
-                }}
-              >
-                添加新类别
-              </Button>
-            </Form.Item>
-          </Col>
-        </Row>
-        <Row gutter={16}>
-          <Col span={12}>
-            <Form.Item
-              name="country"
-              label="国别"
-              labelAlign="left"
-              labelCol={{ span: 4 }}
-            >
-              <Select style={{ width: 120 }}>
-                {InforImport.country?.map((item) => {
-                  return <Option value={item.country}>{item.country}</Option>;
-                })}
-              </Select>
-            </Form.Item>
-          </Col>
-          <Col span={12}>
-            <Form.Item
-              name="name"
-              label="目标舰号或名称"
-              rules={[
-                {
-                  required: type === 1 ? true : false,
-                  message: '请输入目标舰号或名称',
-                },
-              ]}
-            >
-              <Input placeholder="名称" />
-            </Form.Item>
-          </Col>
-        </Row>
-        <Row gutter={16}>
-          <Col span={12}>
-            <Form.Item
-              name="principal_machine"
-              label="主机"
-              labelAlign="left"
-              labelCol={{ span: 4 }}
-            >
-              <Input />
-            </Form.Item>
-          </Col>
-          <Col span={12}>
-            <Form.Item
-              name="auxiliary_machine"
-              label="辅机"
-              labelAlign="left"
-              labelCol={{ span: 4 }}
-            >
-              <Input />
-            </Form.Item>
-          </Col>
-        </Row>
-
-        <Modal
-          visible={visible}
-          onCancel={() => {
-            setVisible(false);
-          }}
-          onOk={() => {
-            form.submit();
-            setVisible(false);
-          }}
-          okText="保存"
-          cancelText="取消"
-          title="添加新类别"
-        >
-          <Form
-            onFinish={(values: any) => {
-              console.log(values);
-              dispatch({
-                type: 'inforImport/addRnType',
-                payload: { name: values.addRnFleet },
-              }).then(() => {});
-            }}
-            form={form}
-          >
-            <Form.Item
-              name="addRnFleet"
-              label="类别名"
-              style={{ marginTop: 20 }}
-            >
-              <Input style={{ width: '80%' }} />
-            </Form.Item>
-          </Form>
-        </Modal>
-      </>
-    );
-  };
-
-  // 目标回声特有信息表单
-  const EchoTarget = () => {
-    const [value, setValue] = useState(-1);
-    const [visible, setVisible] = useState(false);
-
-    const [form] = Form.useForm();
-
-    const onChange = (e) => {
-      console.log('EchoTarget checked', e.target);
-      setValue(e.target.value);
-      if (e.target.value === '添加新类别') {
-        setVisible(true);
-      }
-    };
-
-    return (
-      <>
-        <Row gutter={16}>
-          <Col span={24}>
-            <Form.Item
-              name="te_type"
-              label="目标类型"
-              labelAlign="left"
-              labelCol={{ span: 2 }}
-            >
-              <Radio.Group onChange={onChange} value={value}>
-                {InforImport.teType?.map((item) => {
-                  return <Radio value={item.te_type}>{item.te_type}</Radio>;
-                })}
-                <Radio
-                  value={'添加新类别'}
-                  style={{
-                    display:
-                      CookieUtil.get('role') === '3' ? 'none' : 'inline-block',
-                  }}
-                >
-                  添加新类别
-                </Radio>
-              </Radio.Group>
-            </Form.Item>
-          </Col>
-        </Row>
-        <Row gutter={16}>
-          <Col span={12}>
-            <Form.Item
-              name="country"
-              label="国别"
-              labelAlign="left"
-              labelCol={{ span: 4 }}
-            >
-              <Select style={{ width: 120 }}>
-                {InforImport.country?.map((item) => {
-                  return <Option value={item.country}>{item.country}</Option>;
-                })}
-              </Select>
-            </Form.Item>
-          </Col>
-          <Col span={12}>
-            <Form.Item
-              name="name"
-              label="目标舰号或名称"
-              rules={[
-                {
-                  required: type === 2 ? true : false,
-                  message: '请输入目标舰号或名称',
-                },
-              ]}
-            >
-              <Input placeholder="名称" />
-            </Form.Item>
-          </Col>
-        </Row>
-
-        <Modal
-          visible={visible}
-          onCancel={() => {
-            setVisible(false);
-          }}
-          onOk={() => {
-            form.submit();
-            setVisible(false);
-          }}
-          okText="保存"
-          cancelText="取消"
-          title="添加新类别"
-        >
-          <Form
-            onFinish={(values: any) => {
-              console.log(values);
-              dispatch({
-                type: 'inforImport/addTeType',
-                payload: { name: values.addEchoFleet },
-              }).then(() => {});
-            }}
-            form={form}
-          >
-            <Form.Item
-              name="addEchoFleet"
-              label="类别名"
-              style={{ marginTop: 20 }}
-            >
-              <Input style={{ width: '80%' }} />
-            </Form.Item>
-          </Form>
-        </Modal>
-      </>
-    );
-  };
-
-  // 主动脉冲特有信息表单
-  const PulseTarget = () => {
-    const [value_1, setValue_1] = useState(1);
-    const [value_2, setValue_2] = useState(1);
-    const [visible_1, setVisible_1] = useState(false);
-    const [visible_2, setVisible_2] = useState(false);
-
-    const [form_1] = Form.useForm();
-    const [form_2] = Form.useForm();
-
-    const onChange_1 = (e) => {
-      console.log('PulseTarget checked', e.target);
-      setValue_1(e.target.value);
-      if (e.target.value === '添加新类别') {
-        setVisible_1(true);
-      }
-    };
-    const onChange_2 = (e) => {
-      console.log('PulseTarget checked', e.target);
-      setValue_2(e.target.value);
-      if (e.target.value === '添加新类别') {
-        setVisible_2(true);
-      }
-    };
-
-    return (
-      <>
-        <Row gutter={16}>
-          <Col span={24}>
-            <Form.Item
-              name="ap_type"
-              label="目标类型"
-              labelAlign="left"
-              labelCol={{ span: 2 }}
-            >
-              <Radio.Group onChange={onChange_2} value={value_2}>
-                {InforImport.apType?.map((item) => {
-                  return <Radio value={item.ap_type}>{item.ap_type}</Radio>;
-                })}
-                <Radio
-                  value={'添加新类别'}
-                  style={{
-                    display:
-                      CookieUtil.get('role') === '3' ? 'none' : 'inline-block',
-                  }}
-                >
-                  添加新类别
-                </Radio>
-              </Radio.Group>
-            </Form.Item>
-          </Col>
-        </Row>
-        <Row gutter={16}>
-          <Col span={24}>
-            <Form.Item name="as_type" label="主动声纳类型">
-              <Radio.Group onChange={onChange_1} value={value_1}>
-                {InforImport.asType?.map((item) => {
-                  return <Radio value={item.as_type}>{item.as_type}</Radio>;
-                })}
-                <Radio
-                  value={'添加新类别'}
-                  style={{
-                    display:
-                      CookieUtil.get('role') === '3' ? 'none' : 'inline-block',
-                  }}
-                >
-                  添加新类别
-                </Radio>
-              </Radio.Group>
-            </Form.Item>
-          </Col>
-        </Row>
-        <Row gutter={16}>
-          <Col span={12}>
-            <Form.Item
-              name="country"
-              label="国别"
-              labelAlign="left"
-              labelCol={{ span: 4 }}
-            >
-              <Select style={{ width: 120 }}>
-                {InforImport.country?.map((item) => {
-                  return <Option value={item.country}>{item.country}</Option>;
-                })}
-              </Select>
-            </Form.Item>
-          </Col>
-          <Col span={12}>
-            <Form.Item
-              name="name"
-              label="目标舰号或名称"
-              rules={[
-                {
-                  required: type === 3 ? true : false,
-                  message: '请输入目标舰号或名称',
-                },
-              ]}
-            >
-              <Input placeholder="名称" />
-            </Form.Item>
-          </Col>
-        </Row>
-
-        <Modal
-          visible={visible_2}
-          onCancel={() => {
-            setVisible_2(false);
-          }}
-          onOk={() => {
-            form_2.submit();
-            setVisible_2(false);
-          }}
-          okText="保存"
-          cancelText="取消"
-          title="添加新类别"
-        >
-          <Form
-            onFinish={(values: any) => {
-              console.log(values);
-              dispatch({
-                type: 'inforImport/addApType',
-                payload: { name: values.addPluseFleet },
-              }).then(() => {});
-            }}
-            form={form_2}
-          >
-            <Form.Item
-              name="addPluseFleet"
-              label="类别名"
-              style={{ marginTop: 20 }}
-            >
-              <Input style={{ width: '80%' }} />
-            </Form.Item>
-          </Form>
-        </Modal>
-
-        <Modal
-          visible={visible_1}
-          onCancel={() => {
-            setVisible_1(false);
-          }}
-          onOk={() => {
-            form_1.submit();
-            setVisible_1(false);
-          }}
-          okText="保存"
-          cancelText="取消"
-          title="添加新类别"
-        >
-          <Form
-            onFinish={(values: any) => {
-              console.log(values);
-              dispatch({
-                type: 'inforImport/addAsType',
-                payload: { name: values.addAsType },
-              }).then(() => {});
-            }}
-            form={form_1}
-          >
-            <Form.Item
-              name="addAsType"
-              label="声呐类型名"
-              style={{ marginTop: 20 }}
-            >
-              <Input style={{ width: '80%' }} />
-            </Form.Item>
-          </Form>
-        </Modal>
-      </>
-    );
-  };
-
-  // 动力装置栏
-  const Powerplant = () => {
-    const [value, setValue] = useState(1);
-    const [visible, setVisible] = useState(false);
-    const [form] = Form.useForm();
-
-    const onChange = (e) => {
-      console.log('Powerplant checked', e.target);
-      setValue(e.target.value);
-      // if (e.target.value === '添加新类别') {
-      //   setVisible(true);
-      // }
-    };
-
-    return (
-      <>
-        <Form.Item
-          name="power_engine"
-          label="动力装置"
-          labelAlign="left"
-          labelCol={{ span: 2 }}
-        >
-          <Radio.Group onChange={onChange} value={value}>
-            {InforImport.powerEngine?.map((item) => {
-              return <Radio value={item.name}>{item.name}</Radio>;
-            })}
-          </Radio.Group>
-          <Button
-            value={'添加新类别'}
-            onClick={() => {
-              setVisible(true);
-            }}
-            style={{
-              display: CookieUtil.get('role') === '3' ? 'none' : 'inline-block',
-            }}
-          >
-            添加新类别
-          </Button>
-        </Form.Item>
-        <Modal
-          visible={visible}
-          onCancel={() => {
-            setVisible(false);
-          }}
-          onOk={() => {
-            form.submit();
-            setVisible(false);
-          }}
-          okText="保存"
-          cancelText="取消"
-          title="添加新类别"
-        >
-          <Form
-            onFinish={(values: any) => {
-              console.log(values);
-              dispatch({
-                type: 'inforImport/addPowerEngine',
-                payload: { name: values.addPower },
-              }).then(() => {});
-            }}
-            form={form}
-          >
-            <Form.Item name="addPower" label="类别名" style={{ marginTop: 20 }}>
-              <Input style={{ width: '80%' }} />
-            </Form.Item>
-          </Form>
-        </Modal>
-      </>
-    );
-  };
-
-  // 螺旋桨
-  const Propeller = () => {
-    const [value_2, setValue_2] = useState(-1);
-    const [visible, setVisible] = useState(false);
-    const [form] = Form.useForm();
-
-    const onChange_2 = (e) => {
-      console.log('Propeller checked', e.target);
-      setValue_2(e.target.value);
-      if (e.target.value === '添加新类别') {
-        setVisible(true);
-      } else {
-        sumForm.setFieldsValue({
-          shaft_count: Number(e.target.value.split('_')[0]),
-        });
-        sumForm.setFieldsValue({
-          blade_count: Number(e.target.value.split('_')[1]),
-        });
-        sumForm.setFieldsValue({
-          rotationl_speed: Number(e.target.value.split('_')[2]),
-        });
-      }
-    };
-
-    return (
-      <>
-        <Row gutter={16}>
-          <Col span={10} style={{ display: 'none' }}>
-            <Form.Item name="shaft_count" label="轴数"></Form.Item>
-            <Form.Item name="blade_count" label="叶数"></Form.Item>
-            <Form.Item name="rotationl_speed" label="转速"></Form.Item>
-          </Col>
-          <Col span={24}>
-            <Form.Item
-              name="shaft_blade_rotationl"
-              label="螺旋桨"
-              labelAlign="left"
-              labelCol={{ span: 2 }}
-            >
-              <Radio.Group onChange={onChange_2} value={value_2}>
-                {InforImport.propeller?.map((item) => {
-                  return (
-                    <Radio
-                      value={`${item.shaft_count}_${item.blade_count}_${item.rotationl_speed}`}
-                    >
-                      {item.shaft_count}轴{item.blade_count}叶
-                      {item.rotationl_speed}转速
-                    </Radio>
-                  );
-                })}
-              </Radio.Group>
-              <Button
-                value={'添加新类别'}
-                onClick={() => setVisible(true)}
-                style={{
-                  display:
-                    CookieUtil.get('role') === '3' ? 'none' : 'inline-block',
-                }}
-              >
-                添加新类别
-              </Button>
-            </Form.Item>
-          </Col>
-        </Row>
-
-        <Modal
-          visible={visible}
-          onCancel={() => {
-            setVisible(false);
-          }}
-          onOk={() => {
-            form.submit();
-            setVisible(false);
-          }}
-          okText="保存"
-          cancelText="取消"
-          title="添加新类别"
-        >
-          <Form
-            onFinish={(values: any) => {
-              console.log(values);
-              dispatch({
-                type: 'inforImport/addPropeller',
-                payload: values,
-              }).then(() => {});
-            }}
-            form={form}
-          >
-            <Form.Item
-              name="shaft_count"
-              label="轴数"
-              style={{ marginTop: 20 }}
-            >
-              <InputNumber style={{ width: '80%' }} />
-            </Form.Item>
-            <Form.Item
-              name="blade_count"
-              label="叶数"
-              style={{ marginTop: 20 }}
-            >
-              <InputNumber style={{ width: '80%' }} />
-            </Form.Item>
-            <Form.Item
-              name="rotationl_speed"
-              label="转速"
-              style={{ marginTop: 20 }}
-            >
-              <InputNumber style={{ width: '80%' }} />
-            </Form.Item>
-          </Form>
-        </Modal>
-      </>
-    );
-  };
-
-  // 通用基础信息表单
-  const SignalInfor = () => {
-    return (
-      <>
-        <Row gutter={16}>
-          <Col span={12}>
-            <Form.Item
-              name="speed"
-              label="航速"
-              labelAlign="left"
-              labelCol={{ span: 4 }}
-            >
-              <InputNumber style={{ width: '100%' }} />
-            </Form.Item>
-          </Col>
-          <Col span={12}>
-            <Form.Item
-              name="distance"
-              label="距离"
-              labelAlign="left"
-              labelCol={{ span: 4 }}
-            >
-              <InputNumber style={{ width: '100%' }} />
-            </Form.Item>
-          </Col>
-        </Row>
-        <Row gutter={16}>
-          <Col span={12}>
-            <Form.Item
-              name="collect_d"
-              label="采集日期"
-              labelAlign="left"
-              labelCol={{ span: 4 }}
-            >
-              <DatePicker style={{ width: '100%' }} />
-            </Form.Item>
-          </Col>
-          <Col span={12}>
-            <Form.Item
-              name="collect_t"
-              label="采集时间"
-              labelAlign="left"
-              labelCol={{ span: 4 }}
-            >
-              <TimePicker style={{ width: '100%' }} />
-            </Form.Item>
-          </Col>
-        </Row>
-        <Row gutter={16}>
-          <Col span={12}>
-            <Form.Item
-              name="collect_platform"
-              label="采集平台"
-              labelAlign="left"
-              labelCol={{ span: 4 }}
-            >
-              <Input />
-            </Form.Item>
-          </Col>
-          <Col span={12}>
-            <Form.Item
-              name="location"
-              label="采集位置"
-              labelAlign="left"
-              labelCol={{ span: 4 }}
-            >
-              <Input />
-            </Form.Item>
-          </Col>
-        </Row>
-        <Row gutter={16}>
-          <Col span={12}>
-            <Form.Item
-              name="task_source"
-              label="采集任务源"
-              labelAlign="left"
-              labelCol={{ span: 4 }}
-            >
-              <Input />
-            </Form.Item>
-          </Col>
-          <Col span={12}>
-            <Form.Item
-              name="depth"
-              label="深度"
-              labelAlign="left"
-              labelCol={{ span: 4 }}
-            >
-              <InputNumber style={{ width: '100%' }} />
-            </Form.Item>
-          </Col>
-        </Row>
-      </>
-    );
-  };
-
-  const modifyNoise = (values: any) => {
+  const _modify = (values: any, type: number) => {
     dispatch({
-      type: 'inforImport/modifyNoise',
+      type: `inforImport/${modifyActionNames[type]}`,
       payload: { id: id, body: values },
     }).then(() => {
       dispatch({
@@ -850,30 +91,6 @@ const AddSound: React.FC<{
       });
     });
   };
-
-  const modifyEcho = (values: any) => {
-    dispatch({
-      type: 'inforImport/modifyEcho',
-      payload: { id: id, body: values },
-    }).then(() => {
-      dispatch({
-        type: 'soundList/fetchSoundList',
-      });
-    });
-  };
-
-  const modifyPulse = (values: any) => {
-    dispatch({
-      type: 'inforImport/modifyPulse',
-      payload: { id: id, body: values },
-    }).then(() => {
-      dispatch({
-        type: 'soundList/fetchSoundList',
-      });
-    });
-  };
-
-  const modify = [modifyNoise, modifyEcho, modifyPulse];
 
   return (
     <div style={{ width: '100%' }}>
@@ -911,7 +128,7 @@ const AddSound: React.FC<{
             }
             delete copy_vals['shaft_blade_rotationl'];
             console.log('目标信息表单值', copy_vals);
-            modify[type - 1](copy_vals);
+            _modify(copy_vals, type - 1);
           }
           sumForm.resetFields();
           setCurrent(current + 1);
@@ -924,7 +141,7 @@ const AddSound: React.FC<{
         <Row gutter={16}>
           <Col span={24}>
             <div style={{ marginTop: 0 }}>
-              <TypeRadio />
+              <TypeRadio settype={settype} type={type} />
             </div>
           </Col>
         </Row>
@@ -950,7 +167,12 @@ const AddSound: React.FC<{
                 }}
                 id="radiation_target_div"
               >
-                <RadiationTarget />
+                <RadiationTarget
+                  dispatch={dispatch}
+                  country={InforImport.country}
+                  rnType={InforImport.rnType}
+                  type={type}
+                />
               </div>
               <div
                 style={{
@@ -959,7 +181,12 @@ const AddSound: React.FC<{
                 }}
                 id="echo_target_div"
               >
-                <EchoTarget />
+                <EchoTarget
+                  dispatch={dispatch}
+                  country={InforImport.country}
+                  teType={InforImport.teType}
+                  type={type}
+                />
               </div>
               <div
                 style={{
@@ -968,7 +195,13 @@ const AddSound: React.FC<{
                 }}
                 id="pulse_target_div"
               >
-                <PulseTarget />
+                <PulseTarget
+                  dispatch={dispatch}
+                  country={InforImport.country}
+                  apType={InforImport.apType}
+                  asType={InforImport.asType}
+                  type={type}
+                />
               </div>
             </div>
           </Col>
@@ -982,7 +215,10 @@ const AddSound: React.FC<{
               }}
               id="powerplant_div"
             >
-              <Powerplant />
+              <Powerplant
+                dispatch={dispatch}
+                powerEngine={InforImport.powerEngine}
+              />
             </div>
           </Col>
         </Row>
@@ -992,7 +228,11 @@ const AddSound: React.FC<{
               style={{ marginTop: 0, display: type === 1 ? 'block' : 'none' }}
               id="propeller_div"
             >
-              <Propeller />
+              <Propeller
+                dispatch={dispatch}
+                propeller={InforImport.propeller}
+                sumForm={sumForm}
+              />
             </div>
           </Col>
         </Row>
@@ -1015,7 +255,7 @@ const AddSound: React.FC<{
                 display: type !== -1 ? 'block' : 'none',
               }}
             >
-              <SignalInfor />
+              <SignalInfo />
             </div>
           </Col>
         </Row>
@@ -1038,7 +278,7 @@ const AddSound: React.FC<{
   );
 };
 
-const mapStateToProps = ({ loading, soundList, inforImport }) => {
+const mapStateToProps = ({ inforImport }) => {
   // console.log(loading)
   return {
     InforImport: inforImport,
